% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FeaturePlot_mod}
\alias{FeaturePlot_mod}
\title{Modified featureplot}
\usage{
FeaturePlot_mod(
  object,
  feat,
  min = NA,
  max = NA,
  palette_color = "blue",
  ncol = length(feat)
)
}
\arguments{
\item{object}{the seurat object to visualize the clusters}

\item{feat}{feature to be visualized (genes, module scores...)}

\item{min}{mininum value to be visualized with colorbar}

\item{max}{maximum value to be visualized with colorbar}

\item{palette_color}{color palette used for coloring spots (default: blue)}

\item{ncol}{column numbers for the plot (number of slides in the column)}
}
\description{
Modified featureplot
}
\keyword{internal}
