% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FindMarkers_mod}
\alias{FindMarkers_mod}
\title{Find markers for the cell type based on Seurat FindAllMarkers}
\usage{
FindMarkers_mod(
  object,
  purpose = "marker",
  group.to.find = "seurat_clusters",
  logfc.threshold = 0.25,
  test.use = "wilcox",
  ident.1 = NULL,
  ident.2 = NULL,
  slot = "data",
  assay = NULL,
  min.pct = 0.1,
  only.pos = FALSE,
  latent.vars = NULL,
  min.cells.group = 3
)
}
\arguments{
\item{object}{Seurat object to find markers}

\item{purpose}{purpose of the function: marker discovery or differentially expressed genes finding}

\item{group.to.find}{identity of the group to find markers within}

\item{others}{refer to Seurat::FindMarkers}
}
\value{
dataframe with average log fold change, percentage of cells detected in each group (ident.1 and ident.2), adjusted p-value and average feature value in each group
}
\description{
Calculate markers for the cell type based on Seurat FindAllMarkers. Instead, the column containing average expression in each cluster was added
}
