% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{frequency_boxplot}
\alias{frequency_boxplot}
\title{Draw the frequency boxplot for spots}
\usage{
frequency_boxplot(
  object,
  grp = NULL,
  slide_title = grp,
  group_of_interest = "orig.ident",
  cluster_name = "seurat_clusters",
  cluster_colors = list("#56ebd3", "#1c5e39", "#1fa198", "#7aed7b", "#36a620", "#cadba5",
    "#33547a", "#24a5f7", "#3337a6", "#d678ef", "#9d0d6c", "#b090d4", "#740ece",
    "#ef3df3", "#69345e", "#829499", "#809b31", "#f8ba7c", "#683c00", "#d9dc22",
    "#992a13", "#ec102f", "#df6e78", "#fa7922", "#ae783e", "#7fdc64", "#6f2b6e",
    "#56cac1", "#1b511d", "#ec9fe7", "#214a65", "#b3d9fa", "#1932bf", "#34f50e"),
  x.axis.title = "Groups",
  y.axis.title = "Frequency",
  legend.title = "Clusters",
  x.axis.title.size = 20,
  y.axis.title.size = 20,
  x.axis.text.size = 12,
  y.axis.text.size = 12,
  x.axis.text.angle = 0,
  legend.title.size = 12,
  legend.text.size = 12,
  vis.freq.text = FALSE,
  freq.text.size = 3.5,
  freq.stats = FALSE
)
}
\arguments{
\item{object}{the seurat object to visualize the clusters}

\item{grp}{assigned group name}

\item{slide_title}{change the title of each slide in plot}

\item{group_of_interest}{group name to visualize frequency of clusters (default: orig.ident)}

\item{cluster_name}{cluster name to evaluate frequency in each group (default: seurat_clusters)}

\item{cluster_colors}{the colors used to visualize each cluster}

\item{x.axis.title}{x axis title, y.axis.title: y axis title}

\item{legend.title}{title of the legend}

\item{x.axis.title.size}{x axis title size, y.axis.title.size: y axis title size}

\item{x.axis.text.size}{x axis text size, y.axis.text.size: y axis text size}

\item{x.axis.text.angle}{x axis text angle (default: 0)}

\item{legend.title.size}{legend title size, legend.text.size: legend text size}

\item{vis.freq.text}{visualize frequency inside of the boxplot}

\item{freq.text.size}{size of the frequency text inside of the boxplot}

\item{freq.stats}{whether to return frequency statistics}
}
\description{
Draw the frequency boxplot for spots
}
\keyword{internal}
