% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{horizontal_flip}
\alias{horizontal_flip}
\title{Horizontally flip the image}
\usage{
horizontal_flip(file_dir_vector)
}
\arguments{
\item{file_dir_vector}{'spatial' directory for the 10X Visium-formatted file}
}
\description{
Horizontally flip the image
}
