% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{jscode}
\alias{jscode}
\title{Internal function to close browser}
\format{
An object of class \code{character} of length 1.
}
\usage{
jscode
}
\description{
Close app when closing browser
}
\keyword{internal}
