% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{NS_Forest_R}
\alias{NS_Forest_R}
\title{R wrapper for NSForest}
\usage{
NS_Forest_R(
  object,
  assay = "RNA",
  slot = "data",
  outdir = "./",
  clusterLabelcolumnHeader = "louvain",
  rfTrees = 1000,
  Median_Expression_Level = 0,
  Genes_to_testing = 6,
  betaValue = 0.5,
  conda.env.name = "NSForest"
)
}
\arguments{
\item{object}{Seurat object for the analysis}

\item{assay}{name of the assay for the analysis}

\item{slot}{name of the data slot for the analysis}

\item{outdir}{file directory to save the model and result}

\item{clusterLabelcolumnHeader}{name of the column where cluster assignments reside.}

\item{rfTrees}{Number of trees}

\item{Median_Expression_Level}{median expression level for removing negative markers}

\item{Genes_to_testing}{How many ranked genes by binary score will be evaluated in permutations by fbeta-score}

\item{betaValue}{Set values for fbeta weighting. 1 is default f-measure.}

\item{conda.env.name}{name of the conda environemt (already installed with Anaconda) to use for CellDART analysis (default: 'CellDART')}
}
\value{
dataframe containing optimal combination of marker genes in each cluster
}
\description{
R wrapper for NSForest
}
\references{
Aevermann B, Zhang Y, Novotny M, Keshk M, Bakken T, Miller J, Hodge R, Lelieveldt B, Lein E, Scheuermann RH. A machine learning method for the discovery of minimum marker gene combinations for cell type identification from single-cell RNA sequencing. Genome Res. 2021 Oct;31(10):1767-1780. doi: 10.1101/gr.275569.121.

https://github.com/JCVenterInstitute/NSForest
}
