% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Simple preprocess data}
\usage{
preprocess_data(
  data_dir,
  grp,
  data_type = "Single-cell",
  filter_nfeature_RNA = NULL,
  filter_percent_mt = NULL,
  reference_index = NULL,
  n_var_features = 2000,
  cluster_dim = 30,
  cluster_resolution = 0.5,
  maximum_size = 6000 * 1024^2
)
}
\description{
Simple preprocess data
}
\keyword{internal}
