% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preprocess_data_integ_rpca}
\alias{preprocess_data_integ_rpca}
\title{Integrate datasets using reciprocal PCA in Seurat}
\usage{
preprocess_data_integ_rpca(
  data_dir,
  grp,
  data_type = "Single-cell",
  filter_nfeature_RNA = NULL,
  filter_percent_mt = NULL,
  reference_index = NULL,
  n_var_features = 2000,
  n_integ_features = 2000,
  integ_dim = 50,
  cluster_dim = 30,
  cluster_resolution = 0.5,
  maximum_size = 6000 * 1024^2
)
}
\arguments{
\item{data_dir}{vector for the directories to load the single-cell data}

\item{grp}{group name to assigned to each single-cell data}

\item{data_type}{type of data to integrate (single-cell or spatial)}

\item{filter_nfeature_RNA}{each element is the lower threshold for total number of detected genes in each single-cell data provided}

\item{filter_percent_mt}{each element is the upper threshold for fraction of mitochondrial genes in each single-cell data provided}

\item{reference_index}{vector for index number of the reference single-cell dataset in grp vector (default: NULL)
\itemize{
\item if input is c(1,5) it means that we will use first to fifth element of group as a reference
\item if input is NULL it means that we will use all pairwise anchors for integration
}}

\item{n_var_features}{number of variable genes to be included in each single-cell dataset (default: 2000)}

\item{n_integ_features}{number of genes to be utilized for integration (default: 2000)}

\item{integ_dim}{PC dimensions to be used for integration with reciprocal PCA (default: 50)}

\item{cluster_dim}{PC dimension to be used for clustering of the spots (default: 30)}

\item{cluster_resolution}{resolution during the clustering process (default: 0.5)}

\item{maximum_size}{maximum size allowed for the analysis (default: 6GB)}
}
\value{
Seurat object with multiple integrated datasets
}
\description{
Integrating single-cell or spatial datasets (Reciprocal PCA and Anchor finding)
}
