% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RidgePlot_mod}
\alias{RidgePlot_mod}
\title{Modified ridgeplot}
\usage{
RidgePlot_mod(
  object,
  feat,
  groups = "seurat_clusters",
  lim = c(-0.5, 6),
  alpha = 0,
  title_size = 20,
  x.axis.title = "Expression",
  y.axis.title = "Groups",
  x.axis.title.size = 15,
  y.axis.title.size = 15,
  x.axis.text.size = 12,
  y.axis.text.size = 12,
  ncol = length(feat)
)
}
\arguments{
\item{object}{the seurat object to visualize the clusters}

\item{feat}{feature to be visualized (genes, module scores...)}

\item{groups}{groups to visualize with ridgeplot with}

\item{lim}{minimum and maximum expression value to be visualized with colorbar}

\item{alpha}{transparency of the ridgeplot (default=0)}

\item{title_size}{size of the title in each plot}

\item{x.axis.title}{x axis title, y.axis.title: y axis title}

\item{x.axis.title.size}{x axis title size, y.axis.title.size: y axis title size}

\item{x.axis.text.size}{x axis text size, y.axis.text.size: y axis text size}

\item{ncol}{column numbers for the plot (number of plots in the column)}
}
\description{
Modified ridgeplot
}
\keyword{internal}
