% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Function to run STqunatools app}
\usage{
run_app()
}
\value{
starts shiny application for STquantools
examples
STquantool::run_app()
}
\description{
ST analysis tool to visualize and quantify multiple datasets
}
