% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_feat_plot_by_n}
\alias{save_feat_plot_by_n}
\title{Automatically saving spatial feature plot by n}
\usage{
save_feat_plot_by_n(
  object,
  feats_of_interest,
  grp = NULL,
  slide_title = grp,
  by_n = 3,
  min = NA,
  max = NA,
  crop = TRUE,
  pt.size.factor = 1.8,
  alpha = c(1, 1),
  image.alpha = 1,
  title_size = 14,
  face = "bold",
  ncol = length(grp),
  palette_color = NULL,
  color_bar_mode = "default",
  color_bar_loc = "bottom",
  slide.to.visualize = NULL,
  cluster_name = "seurat_clusters",
  spot.cluster.highlight = NULL,
  height = 8,
  width = 8,
  dpi = 200,
  save_path = "./figures/integ",
  file_name = "feats",
  sort.by.expression = TRUE,
  assay = NULL,
  slot = "data"
)
}
\arguments{
\item{object}{the seurat object to visualize the clusters}

\item{feats_of_interest}{feature to be visualized (genes, module scores...)}

\item{grp}{assigned group name}

\item{slide_title}{change the group title in each plot}

\item{by_n}{the number of the features to be visualized in each plot}

\item{min}{mininum value to be visualized with colorbar}

\item{max}{maximum value to be visualized with colorbar}

\item{crop}{whether to crop the image: default = TRUE}

\item{pt.size.factor}{scale the size of the plot: default: 1.6}

\item{alpha}{transparency of the color spot on the tissue: default: c(0,1)}

\item{image.alpha}{transparency of the background tissue: default: 1}

\item{title_size}{size of the title for each spatial plot}

\item{face}{the type of the text on the plot}

\item{ncol}{column numbers for the plot (number of slides in the column)}

\item{palette_color}{color palette used for coloring spots (default: Seurat colormap)}

\item{color_bar_mode}{change the mode of colorbar visualization (default: 'default')
\itemize{
\item set color_bar_mode = 'combined' to visualize colorbar only once
}}

\item{color_bar_loc}{location of the colorbar: bottom, top, left, right}

\item{slide.to.visualize}{vector representing number of slide to visualize: eg: c(1,3)}

\item{cluster_name}{group name of a specific spot cluster to be visualized}

\item{spot.cluster.highlight}{vector of spot cluster name to be highlighted}

\item{height}{height of the printed image (by inch)}

\item{width}{width of the printed image (by inch)}

\item{dpi}{dpi of the image to be saved}

\item{save_path}{the path to save the file}

\item{file_name}{header name of the image files to save}

\item{sort.by.expression}{sort the given feature according the the average expression in whole tissue}

\item{assay}{assay of the data to use for calculation of average expression}

\item{slot}{name of the slot to use in the given Seurat object}
}
\description{
Automatically saving spatial feature plot by n
}
