% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_files_Modal}
\alias{save_files_Modal}
\title{Modal to save the rds files}
\usage{
save_files_Modal(
  text = "single-cell",
  text_for_purpose = "Will you save the ",
  text_for_add = " data?",
  text_input_name = "sc_save_name",
  text_input_explain = "Name of the file",
  file_save_name = "sc_data",
  add_text_input = FALSE,
  add_text_input_name = "sc_save_name_add",
  add_text_input_explain = "Name of the file",
  add_file_save_name = "sc_data",
  action_button_name = "ok_sc"
)
}
\description{
Modal to save the rds files
}
\keyword{internal}
