% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_gene_list_Modal}
\alias{save_gene_list_Modal}
\title{Modal to save gene list}
\usage{
save_gene_list_Modal(
  text_for_purpose = "Will you save the given gene list?",
  text_input_name = "gene_save_name",
  text_input_explain = "Name of the gene list",
  file_save_name = "Gene list",
  action_button_name = "ok_gene"
)
}
\description{
Modal to load csv or txt files
}
\keyword{internal}
