% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_plot_Modal}
\alias{save_plot_Modal}
\title{Save the images}
\usage{
save_plot_Modal(
  slider_input_name = "save_dpi",
  action_button_name = "save_start"
)
}
\description{
Save the images
}
\keyword{internal}
