% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_plot_wh_Modal}
\alias{save_plot_wh_Modal}
\title{Modal to save the deg output plot files}
\usage{
save_plot_wh_Modal(
  slider_input_name = "save_dpi",
  width_name = "deg_width",
  height_name = "deg_height",
  width_value = 13,
  height_value = 15,
  action_button_name = "save_start"
)
}
\description{
Save the deg output plot files
}
\keyword{internal}
