% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{spatial_cluster_plot}
\alias{spatial_cluster_plot}
\title{Visualizing the spot clusters: spatial mapping with labels}
\usage{
spatial_cluster_plot(
  object,
  grp = NULL,
  slide_title = grp,
  cluster_name = "seurat_clusters",
  crop = TRUE,
  alpha = c(1, 1),
  image.alpha = 0.6,
  pt.size.factor = 1.6,
  title_size = 14,
  face = "bold",
  label = TRUE,
  label.size = 3,
  ncol = length(grp),
  slide.to.visualize = NULL,
  spot.cluster.highlight = NULL
)
}
\arguments{
\item{object}{the seurat object to visualize the clusters}

\item{grp}{assigned group name}

\item{slide_title}{change the title of each slide in plot}

\item{cluster_name}{name of the clusters to visualize}

\item{crop}{whether to crop the image: default = TRUE}

\item{alpha}{transparency of the color spot on the tissue: default: c(0,1)}

\item{image.alpha}{transparency of the tissue image (default 0.6)}

\item{pt.size.factor}{size of the spot visualized on the tissue}

\item{title_size}{size of the title for each spatial plot}

\item{face}{the type of the text on the plot}

\item{label}{whether to label the spot cluster}

\item{label.size}{size of the text on the plot}

\item{ncol}{column numbers for the plot (number of slides in the column)}

\item{slide.to.visualize}{vector representing number of slide to visualize: eg: c(1,3)}

\item{spot.cluster.highlight}{vector of spot cluster name to be highlighted}
}
\description{
Visualizing the spot clusters: spatial mapping with labels
}
