% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{spatial_feat_plot_groups}
\alias{spatial_feat_plot_groups}
\title{Visualizing spatial gene expression}
\usage{
spatial_feat_plot_groups(
  object,
  feat,
  grp = NULL,
  slide_title = grp,
  min = NA,
  max = NA,
  crop = TRUE,
  pt.size.factor = 1.8,
  alpha = c(1, 1),
  image.alpha = 1,
  title_size = 14,
  face = "bold",
  ncol = length(grp),
  palette_color = NULL,
  color_bar_mode = "default",
  color_bar_loc = "bottom",
  slide.to.visualize = NULL,
  cluster_name = "seurat_clusters",
  spot.cluster.highlight = NULL
)
}
\arguments{
\item{object}{the seurat object to visualize the clusters}

\item{feat}{feature to be visualized (genes, module scores...)}

\item{grp}{assigned group name}

\item{slide_title}{change the title of each slide in plot}

\item{min}{mininum value to be visualized with colorbar: default: NA}

\item{max}{maximum value to be visualized with colorbar: default: NA}

\item{crop}{whether to crop the image: default = TRUE}

\item{pt.size.factor}{scale the size of the plot: default: 1.8}

\item{alpha}{transparency of the color spot on the tissue: default: c(0,1)}

\item{image.alpha}{transparency of the background tissue: default: 1}

\item{title_size}{size of the title for each spatial plot}

\item{face}{the type of the text on the plot}

\item{ncol}{column numbers for the plot (number of slides in the column)}

\item{palette_color}{color palette used for coloring spots (default: Seurat colormap)}

\item{color_bar_mode}{change the mode of colorbar visualization (default: 'default')
#' \itemize{
\item set color_bar_mode = 'combined' to visualize colorbar only once
}}

\item{color_bar_loc}{location of the colorbar: bottom, top, left, right}

\item{slide.to.visualize}{vector representing number of slide to visualize: eg: c(1,3)}

\item{cluster_name}{group name of a specific spot cluster to be visualized}

\item{spot.cluster.highlight}{vector of spot cluster name to be highlighted}
}
\description{
Visualizing spatial gene expression
}
\references{
https://wilkelab.org/cowplot/articles/shared_legends.html
}
