% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vertical_flip}
\alias{vertical_flip}
\title{Vertically flip the image}
\usage{
vertical_flip(file_dir_vector)
}
\arguments{
\item{file_dir_vector}{'spatial' directory for the 10X Visium-formatted file}
}
\description{
Vertically flip the image
}
