% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{NS_Forest_R}
\alias{NS_Forest_R}
\title{R wrapper for NSForest}
\usage{
NS_Forest_R(
  object,
  assay = "RNA",
  slot = "data",
  NSForest_dir = "./",
  outdir = "./",
  clusterLabelcolumnHeader = "louvain",
  rfTrees = 1000,
  Median_Expression_Level = 0,
  Genes_to_testing = 6,
  betaValue = 0.5,
  conda.env.name = "NSForest"
)
}
\arguments{
\item{object}{Seurat object for the analysis}

\item{assay}{name of the assay for the analysis}

\item{slot}{name of the data slot for the analysis}

\item{outdir}{file directory to save the model and result}

\item{clusterLabelcolumnHeader}{name of the column where cluster assignments reside.}

\item{rfTrees}{Number of trees}

\item{Median_Expression_Level}{median expression level for removing negative markers}

\item{Genes_to_testing}{How many ranked genes by binary score will be evaluated in permutations by fbeta-score}

\item{betaValue}{Set values for fbeta weighting. 1 is default f-measure.}

\item{NSForest.dir}{R wrapper python file for the NSForest analysis}

\item{conda.evn.name}{name of the conda environemt (already installed with Anaconda) to use for CellDART analysis (default: 'CellDART')}
}
\description{
R wrapper for NSForest
}
\keyword{internal}
