% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pred_cellf_celldart}
\alias{pred_cellf_celldart}
\title{Predict spatial cell composition}
\usage{
pred_cellf_celldart(
  sp_data,
  sc_data,
  outdir = ".",
  sp_subset = FALSE,
  spot.cluster.name = "seurat_clusters",
  spot.cluster.of.interest = NULL,
  conda.env.name = "CellDART",
  gpu = TRUE,
  metadata_celltype = "celltype",
  num_markers = 20,
  seed_num = 0,
  nmix = 8,
  npseudo = 20000,
  alpha = 0.6,
  alpha_lr = 5,
  emb_dim = 64,
  batch_size = 512,
  n_iterations = 3000,
  init_train_epoch = 10
)
}
\description{
Predict spatial cell composition using CellDART
}
\keyword{internal}
