% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{quantitation_plot}
\alias{quantitation_plot}
\title{Quantitation of the data}
\usage{
quantitation_plot(
  object,
  data.to.use = NULL,
  group.to.compare = "orig.ident",
  group.to.facet = "seurat_clusters",
  agg.to.boxplot = FALSE,
  group.to.split = "orig.ident",
  group.to.recode = NULL,
  calculate_mode = "mean",
  pairwise.comp.stats = NULL,
  cluster_colors = "npg",
  plot_ncol = 4,
  x.axis.title = "",
  y.axis.title = "cell fraction",
  x.axis.title.size = 20,
  y.axis.title.size = 20,
  x.axis.text.size = 10,
  y.axis.text.size = 12,
  x.axis.text.angle = 90,
  legend.title.size = 12,
  legend.text.size = 12,
  vis.cellf.text = FALSE,
  cellf.text.size = 3.5,
  cellf.stats = FALSE,
  spot.total.num.stats = FALSE
)
}
\description{
Quantitation of the data
}
\keyword{internal}
