% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recluster_dataset_rpca}
\alias{recluster_dataset_rpca}
\title{Recluster the dataset}
\usage{
recluster_dataset_rpca(
  data,
  split.by = "orig.ident",
  data_type = "Single-cell",
  n_var_features = 2000,
  n_integ_features = 2000,
  integ_dim = 50,
  cluster_dim = 30,
  cluster_resolution = 0.5,
  maximum_size = 6000 * 1024^2
)
}
\arguments{
\item{data}{single-cell or sptial data to recluster}

\item{data_type}{type of data to integrate (single-cell or spatial)}

\item{n_var_features}{number of variable genes to be included in each single-cell dataset (default: 2000)}

\item{n_integ_features}{number of genes to be utilized for integration (default: 2000)}

\item{integ_dim}{PC dimensions to be used for integration with reciprocal PCA (default: 50)}

\item{cluster_dim}{PC dimension to be used for clustering of the spots (default: 30)}

\item{cluster_resolution}{resolution during the clustering process (default: 0.5)}

\item{maximum_size}{maximum size allowed for the analysis (default: 6GB)}
}
\description{
Recluster the dataset
}
\keyword{internal}
