% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Function to run STquantools app}
\usage{
run_app()
}
\value{
Starts shiny application for STquantools
}
\description{
ST analysis tool to visualize and quantify multiple datasets
}
\examples{
STquantool::run_app()
}
